Implement Crawling Function

+++ src/main/java/com/toyproject/abs_explorer/RankRepository.java
@@ -2,5 +2,5 @@ package com.toyproject.abs_explorer;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RankRepository extends JpaRepository<Book, Integer>{
+public interface RankRepository extends JpaRepository<Book, Long>{
 }
